finished

+++ src/Components/TablaContactos.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import Swal from "sweetalert2";
 import FormularioAdd from "./FormularioAdd";
 
 const TablaContactos = ({
@@ -12,7 +13,20 @@ const TablaContactos = ({
       type: "delete",
       payload: id,
     };
-    dispatch(actionDelete);
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(actionDelete);
+      }
+    });
   };
 
   return (
